Add the ability to signal daemon to start/stop service.
Since certain tasks like start/stop are not instantaneous, we offload the heavy
lifting to the daemon which can start and stop the processes. To do that, we
create a new channel and then we use that Channel to signal the server with the
unit's name.

Server then looks up the name and performs the action.